Fix Guake mode not working on X11 when opening a window
When the Guake mode is enabled, opening a Terminus window in X11 disables
the hotkey for showing the Guake terminal.

This patch fixes this by ensuring that there is only one instance of
Terminus, and other instances opened will use DBus to ask that instance
to open a new window, or show the Guake terminal.