Added compatibility with 1.14.1 extension
If the package is updated and terminus is killed, the extension
tries to launch a new terminus with --check_guake_wayland or
--check_guake_x11; but those parameters aren't recongnized
anymore, so it fails. Unfortunately, the old extension tries
to reload it over and over again, and since the timeout is
100ms, it is too fast and Gnome Shell hangs.

After rebooting, everything will work fine, but it isn't a
desirable situation.

This patch adds unofficial support for those parameters, thus
avoiding this failure during the update.