Use Meta.WaylandClient in the extension
Previously, the extension launched Terminus using a normal
Gio.Subprocess call. This has several problems:

* it requires a trick to detect that this is the Guake window
  in Wayland
* inside Gnome Shell, the open file descriptors limit is set
  to the hard max (about a million), which is inherited by
  the program (Terminus) launched by Gio.Subprocess.

The solution is to use Meta.WaylandClient, which offers a
simple way of detecting if a window belongs to a process. And
it also resets the open file descriptors limit to the original
soft value before handing over the control to the Terminus
code. This fixes several problems, like Valgrind not working or
Steam hanging up during launch (unless an 'ulimit -n 1024' is
manually done in Terminus before running them).