Fix crash when closing/moving terminals
When a terminal is moved (with DnD) or closed, the terminal tree
must be re-balanced. Unfortunately, there was a bug that
prevented the upper container to be correctly updated in some
cases, so, when moving the focus with the keyboard, it could
result in a crash.

This patch fixes it.